FogBugz #582382: Fix integer overflow for Arria 10
dramaddrw data structure are defined as u32. Compiler assume 32-bit
arithmetic on expression and the shifting operation to be done on a
32-bit value. So, the expression would overflow if the shifting
is more than 32-bit.

+++ arch/arm/cpu/armv7/socfpga_arria10/sdram.c
@@ -409,13 +409,15 @@ unsigned long long sdram_size_calc(void)
 	union dramaddrw_reg dramaddrw =
 		(union dramaddrw_reg)readl(&socfpga_io48_mmr_base->dramaddrw);
 
-	unsigned long long size = (1 << (dramaddrw.cfg_cs_addr_width +
+	unsigned long long size = ((unsigned long long)1 <<
+				(dramaddrw.cfg_cs_addr_width +
 				dramaddrw.cfg_bank_group_addr_width +
 				dramaddrw.cfg_bank_addr_width +
 				dramaddrw.cfg_row_addr_width +
 				dramaddrw.cfg_col_addr_width));
 
-	size *= (2 << (readl(&socfpga_ecc_hmc_base->ddrioctrl) &
+	size *= ((unsigned long long)2 <<
+		       (readl(&socfpga_ecc_hmc_base->ddrioctrl) &
 		       ALT_ECC_HMC_OCP_DDRIOCTRL_IO_SIZE_MSK));
 
 	return size;
